Update lib.php
The changes are in the elses in the offlinequiz_extend_settings_navigation function, starting line 1327: It's not a bug, strickly speaking, that renders the mod unusable but if you're on a boost theme (parent or child) and if you're on debug mode you get errors on certain pages as the course parameters. This is caused by the navigation difference between boost and classic themes.
To prevent this, The $url variable in the elses first check that $PAGE->cm->id is set (ternary). If it is then everything is fine and dandy. If not, then we still affect the $PAGE->url
I guess it'd been possible to juste add "&& !isset($PAGE->cm->id)" to the if declaration but it seemed less readable for someone not familiar with the code.